+++ config_test.go
@@ -55,7 +55,8 @@ func TestConfigNewHandlerOptions(t *testing.T) {
 
 // go test -v -cover -count=1 -test.cpu=1 -run=^TestConfigNewHandler$
 func TestConfigNewHandler(t *testing.T) {
-	RegisterHandler(t.Name(), func(w io.Writer, opts *slog.HandlerOptions) slog.Handler {
+	handlerName := t.Name()
+	RegisterHandler(handlerName, func(w io.Writer, opts *slog.HandlerOptions) slog.Handler {
 		return &testConfigHandler{
 			w:    w,
 			opts: *opts,
@@ -67,7 +68,7 @@ func TestConfigNewHandler(t *testing.T) {
 
 	conf := &config{
 		level:       slog.LevelWarn,
-		handler:     t.Name(),
+		handler:     handlerName,
 		newWriter:   newWriter,
 		replaceAttr: replaceAttr,
 		withSource:  true,