+++ doc.go
@@ -299,5 +299,5 @@ package logit // import "github.com/FishGoddess/logit"
 
 const (
 	// Version is the version string representation of logit.
-	Version = "v1.5.1-alpha"
+	Version = "v1.5.2-alpha"
 )