+++ handler_test.go
@@ -21,14 +21,16 @@ import (
 	"testing"
 )
 
-// go test -v -cover -count=1 -test.cpu=1 -run=^TestPickHandler$
-func TestPickHandler(t *testing.T) {
+// go test -v -cover -count=1 -test.cpu=1 -run=^TestGetHandlerFunc$
+func TestGetHandlerFunc(t *testing.T) {
 	newHandler := func(w io.Writer, opts *slog.HandlerOptions) slog.Handler {
 		return nil
 	}
 
-	newHandlers[t.Name()] = newHandler
-	got, err := PickHandler(t.Name())
+	handler := t.Name()
+	newHandlers[handler] = newHandler
+
+	got, err := getHandlerFunc(handler)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -40,23 +42,25 @@ func TestPickHandler(t *testing.T) {
 
 // go test -v -cover -count=1 -test.cpu=1 -run=^TestRegisterHandler$
 func TestRegisterHandler(t *testing.T) {
-	if err := RegisterHandler("text", nil); err == nil {
-		t.Fatal("register an existed handler func should be failed")
+	for name := range newHandlers {
+		if err := RegisterHandler(name, nil); err == nil {
+			t.Fatal("register an existed handler func should be failed")
+		}
 	}
 
-	handler := "new"
-	newHandlerFunc := func(w io.Writer, opts *slog.HandlerOptions) slog.Handler { return nil }
+	handler := t.Name()
+	newHandler := func(w io.Writer, opts *slog.HandlerOptions) slog.Handler { return nil }
 
-	if err := RegisterHandler(handler, newHandlerFunc); err != nil {
+	if err := RegisterHandler(handler, newHandler); err != nil {
 		t.Fatal(err)
 	}
 
-	newHandler, ok := newHandlers[handler]
+	got, ok := newHandlers[handler]
 	if !ok {
 		t.Fatalf("handler %s not found", handler)
 	}
 
-	if fmt.Sprintf("%p", newHandler) != fmt.Sprintf("%p", newHandlerFunc) {
+	if fmt.Sprintf("%p", got) != fmt.Sprintf("%p", newHandler) {
 		t.Fatal("newHandler registered is wrong")
 	}
 }